* Implemented Services * Added spring docs swagger api
