--- src/main/java/com/example/socialboxbackend/service/GroupService.java
@@ -1,4 +1,0 @@
-package com.example.socialboxbackend.service;
-
-public interface GroupService {
-}
