--- src/main/java/com/example/socialboxbackend/service/UserService.java
@@ -1,4 +1,0 @@
-package com.example.socialboxbackend.service;
-
-public interface UserService {
-}
