--- src/main/java/com/example/socialboxbackend/service/impl/MovieServiceImpl.java
@@ -1,4 +1,0 @@
-package com.example.socialboxbackend.service.impl;
-
-public class MovieServiceImpl {
-}
