+++ src/main/java/com/socialbox/model/Group.java
@@ -1,4 +1,4 @@
-package com.example.socialboxbackend.model;
+package com.socialbox.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,17 +6,17 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 
+import java.util.HashSet;
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class MovieList {
+public class Group {
     @Id
-    public String listId;
-    public String linkedUserId;
-    public String linkedGroupId;
-    public List<String> moviesId;
-    public List<String> sharedUserIds;
+    public String groupId;
+    public String groupName;
+    public List<String> groupMovieList;
+    public HashSet<String> usersId;
 }
