+++ src/main/java/com/socialbox/model/Movie.java
@@ -1,4 +1,4 @@
-package com.example.socialboxbackend.model;
+package com.socialbox.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,17 +6,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 
-import java.util.HashSet;
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Group {
+public class Movie {
     @Id
-    public String groupId;
-    public String groupName;
-    public List<String> groupMovieList;
-    public HashSet<String> usersId;
+    public String movieId;
+    public String movieName;
+    public double movieRating;
+    public List<String> movieReviews;
 }
