+++ src/main/java/com/socialbox/repository/MovieListRepository.java
@@ -1,6 +1,6 @@
-package com.example.socialboxbackend.repository;
+package com.socialbox.repository;
 
-import com.example.socialboxbackend.model.MovieList;
+import com.socialbox.model.MovieList;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface MovieListRepository extends MongoRepository<MovieList, String> {
