+++ src/main/java/com/socialbox/repository/SocialBoxRepository.java
@@ -1,6 +1,6 @@
-package com.example.socialboxbackend.repository;
+package com.socialbox.repository;
 
-import com.example.socialboxbackend.model.SocialBox;
+import com.socialbox.model.SocialBox;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface SocialBoxRepository extends MongoRepository<SocialBox, String> {
