+++ src/main/java/com/socialbox/service/GroupService.java
@@ -1,0 +1,4 @@
+package com.socialbox.service;
+
+public interface GroupService {
+}
