+++ src/main/java/com/socialbox/service/UserService.java
@@ -1,0 +1,4 @@
+package com.socialbox.service;
+
+public interface UserService {
+}
