+++ src/main/java/com/socialbox/service/impl/GroupServiceImpl.java
@@ -1,0 +1,6 @@
+package com.socialbox.service.impl;
+
+import com.socialbox.service.GroupService;
+
+public class GroupServiceImpl implements GroupService {
+}
