+++ src/main/java/com/socialbox/service/impl/MovieListServiceImpl.java
@@ -1,0 +1,6 @@
+package com.socialbox.service.impl;
+
+import com.socialbox.service.MovieListService;
+
+public class MovieListServiceImpl implements MovieListService {
+}
