+++ src/main/java/com/socialbox/service/impl/MovieServiceImpl.java
@@ -1,0 +1,6 @@
+package com.socialbox.service.impl;
+
+import com.socialbox.service.MovieService;
+
+public class MovieServiceImpl implements MovieService {
+}
