+++ src/main/java/com/socialbox/service/impl/UserServiceImpl.java
@@ -1,0 +1,6 @@
+package com.socialbox.service.impl;
+
+import com.socialbox.service.UserService;
+
+public class UserServiceImpl implements UserService {
+}
