+++ src/test/java/com/example/socialboxbackend/SocialBoxApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SocialBoxBackendApplicationTests {
+class SocialBoxApplicationTests {
 
     @Test
     void contextLoads() {
